move out snapshot system
